backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlalchemy-utils
psycopg2-binary
passlib[bcrypt]
bcrypt<4
pydantic
pydantic[email]
python-dotenv
chardet<6
greenlet
typing_extensions
python-jose[cryptography]
firebase-admin
python-multipart
paddleocr
pillow
numpy
pypdfium2
fastapi_mail